pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
ipython = "^8.26.0"
djangorestframework-simplejwt = "^5.3.1"
coreapi = "^2.3.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
